Move exception handlers to separate module
